server/requirements.txt
websockets
paho-mqtt
vosk
dotenv